Move connect root retrieval and cert signing logic out of the RPC endpoints
The code now lives on the Server type itself. This was done so that all of this could be shared with auto config certificate signing.
